requirements.txt
streamlit
PyPDF2
python-dotenv
transformers
torch
langchain
langchain-community